Rename "quantum" subfolder to "sdp" in preperation for eventual LP sister class.
