+++ inflation/sdp/monomial_classes.py
