+++ test/test_monomial_properties.py
@@ -2,8 +2,8 @@ import unittest
 import numpy as np
 
 from inflation import InflationProblem, InflationSDP
-from inflation.quantum.fast_npa import nb_is_knowable as is_knowable
-from inflation.quantum.fast_npa import nb_is_physical as is_physical
+from inflation.sdp.fast_npa import nb_is_knowable as is_knowable
+from inflation.sdp.fast_npa import nb_is_physical as is_physical
 
 
 scenario = InflationProblem({"lambda": ["A", "B"],
